Move concatenation helpers into new module
